requirements.txt
fluent~=0.10.0
pydash~=8.0.1
chardet~=5.2.0
deep-translator~=1.11.4
python-telegram-bot~=22.0
GitPython~=3.1.44